Unify doc extensions and introduce new deployment config to conditionally show documentation links
