+++ apps/console/java/org.wso2.identity.apps.console.server.feature/resources/deployment.config.json.j2
@@ -238,6 +238,9 @@
         {% if console.ui.show_app_switch_button is defined %}
         "showAppSwitchButton": {{ console.ui.show_app_switch_button }},
         {% endif %}
+        {% if console.ui.show_doc_links is defined %}
+        "showDocLinks": {{ console.ui.show_doc_links }},
+        {% endif %}
         {% if identity_mgt.user_claim_update.enable_multiple_emails_and_mobile_numbers is defined %}
         "isMultipleEmailsAndMobileNumbersEnabled": {{ identity_mgt.user_claim_update.enable_multiple_emails_and_mobile_numbers }},
         {% endif %}
