+++ modules/react-components/src/components/documentation/documentation-provider.tsx
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2021, WSO2 LLC. (https://www.wso2.com). All Rights Reserved.
+ * Copyright (c) 2021-2026, WSO2 LLC. (https://www.wso2.com).
  *
  * WSO2 LLC. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -27,6 +27,10 @@ interface DocumentationProviderPropsInterface<T> {
      * Initial reducer state.
      */
     links?: T;
+    /**
+     * Whether to show documentation links.
+     */
+    showDocLinks?: boolean;
 }
 
 /**
@@ -35,20 +39,21 @@ interface DocumentationProviderPropsInterface<T> {
  * @param props - Wrap content/elements.
  * @returns DocumentationContext Provider.
  */
-export const DocumentationProvider = <T extends unknown>(
+export const DocumentationProvider = <T,>(
     props: PropsWithChildren<DocumentationProviderPropsInterface<T>>
 ): ReactElement => {
 
     const {
         children,
-        links
+        links,
+        showDocLinks = false
     } = props;
 
     /**
      * Render state, dispatch and special case actions.
      */
     return (
-        <DocumentationContext.Provider value={ { links } }>
+        <DocumentationContext.Provider value={ { links, showDocLinks } }>
             { children }
         </DocumentationContext.Provider>
     );
