+++ modules/react-components/src/components/documentation/use-documentation.tsx
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2021, WSO2 LLC. (https://www.wso2.com). All Rights Reserved.
+ * Copyright (c) 2021-2026, WSO2 LLC. (https://www.wso2.com).
  *
  * WSO2 LLC. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -27,7 +27,13 @@ export const useDocumentation = ():{ getLink: (key: string) => string } => {
 
     const config = useContext(DocumentationContext);
 
-    const getLink = (key: string):string => get(config?.links, key);
+    const getLink = (key: string):string => {
+        if (config?.showDocLinks === false) {
+            return undefined;
+        }
+
+        return get(config?.links, key);
+    };
 
     return { getLink };
 };
